@@ -18,9 +18,9 @@
     "source": "https://github.com/gbuckingham89/laraflash"
   },
   "require": {
-    "php": "^7.1|^8.0",
-    "illuminate/session": "^5.6|^5.7|^5.8|^6.0|^7.0|^8.0",
-    "illuminate/support": "^5.6|^5.7|^5.8|^6.0|^7.0|^8.0"
+    "php": "^7.4|^8.0",
+    "illuminate/session": "^7.0|^8.0",
+    "illuminate/support": "^7.0|^8.0"
   },
   "autoload": {
     "psr-4": {
